+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Add correct aria-label attribute to "Patron look-up" button. Refs UICHKOUT-934.
 * Implement support for fields without information. Refs UICHKOUT-945.
 * Check out items using new endpoint from mod-circulation-bff. Refs UICHKOUT-946.
+* Use `history.push` instead of `mutator.query` updating. Refs UICHKOUT-948.
 
 ## [11.0.2] (https://github.com/folio-org/ui-checkout/tree/v11.0.2) (2024-11-30)
 [Full Changelog](https://github.com/folio-org/ui-checkout/compare/v11.0.1...v11.0.2)